Get the original example working in the function component framework.
All abilities for Amalgate have been broken down to basic functionality: given the lines that came before a method, the arguments, (and the lines in the body for multi-line), each function can implement themselves in very encapsulated, modular ways. The reader simply looks for the pattern of an amalgate function, rather than an actual function name. To add a method to amalgate, simply make a file that implements AmalgateFunction, and add an instance of it to the AmalgateReader.